# openapi/openapi.spec.yaml
openapi: "3.0.0"
paths:
  /profile:
    get:
      operationId: "getProfile"
      parameters: []
      responses:
        200:
          description: ""
  /users:
    get:
      operationId: "usersFindAll"
      parameters: []
      responses:
        200:
          description: ""
    post:
      operationId: "usersCreate"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        201:
          description: ""
  /users/{id}:
    get:
      operationId: "usersFindOne"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
  /ship:
    get:
      operationId: "shipsFindAll"
      parameters: []
      responses:
        200:
          description: ""
    post:
      operationId: "shipsCreate"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipDto"
      responses:
        201:
          description: ""
  /ship/{id}:
    get:
      operationId: "shipsFindOne"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
    delete:
      operationId: "shipsDeleteOne"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
    put:
      operationId: "shipsUpdateOne"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipDto"
      responses:
        200:
          description: ""
info:
  title: "AquilaCode API"
  description: "AquilaCode API, for controlling the universe"
  version: "1.0"
  contact: {}
tags:
  - name: "AquilaCode"
    description: ""
servers: []
components:
  schemas:
    CreateUserDto:
      type: "object"
      properties: {}
    CreateShipDto:
      type: "object"
      properties: {}